refactor(api/ai): split the run-profile finalizers out of runLoop.ts (#4451)
`runLoop.ts` had grown to 2795 lines, carrying the SDK loop, the pre/post
tool-use hooks, the run-context loader, the finish path AND the four
per-profile finalizers. This is the split the issue asked for.

Pure move — no behaviour change:

- `runFinalizers.ts` (new) takes `finalizeVerdict`, `finalizeSweep`,
  `finalizeNarrative` and `finalizeTicketTriage` verbatim, plus
  `isRunStillRunning`, which was used exclusively by those four.
- `runLoopTypes.ts` (new) takes the shape contracts both sides need:
  `RunRow`, `AgentRow`, `RunContext`, `LoopResult`, `AgentRunOutcome`,
  `OutcomeProposedAction`, `OutcomeExecutedAction`, `TicketProposalOutcome`.

The types file exists so the graph stays strictly acyclic — `runLoop` →
`runFinalizers` → `runLoopTypes`, with `runLoop` → `runLoopTypes` as well.
Leaving `RunContext`/`LoopResult` in `runLoop.ts` would have forced
`runFinalizers.ts` to import back from its own caller; that cycle would be
type-only and erased at runtime, but it is exactly the direction the split
is meant to remove.

Two deliberate, small non-moves:

- `runLoop.ts` re-exports the four outcome types so every existing importer
  (`runTrace.ts`, `fixWatch.test.ts`, `aiAgentSdkTools.test.ts`,
  `redTeam.contract.test.ts`, the four `runLoop*.test.ts` suites) keeps its
  current import path. No test file changed.
- `inSystemDbContext` is a local copy in `runFinalizers.ts` rather than an
  export from `runLoop.ts` — importing it back would re-introduce the cycle.
  `runService.ts` and `effectivePolicy.ts` already carry their own copies of
  this four-line wrapper; deduplicating all of them is a separate change.

`runLoop.ts` 2795 → 2098 lines. Every moved block is byte-identical to the
original apart from the added `export` keywords (verified by diffing the
extracted ranges).

Verified: `tsc --noEmit` clean; `eslint` clean on all three files; the four
`runLoop*.test.ts` suites plus `aiAgentSdkTools`, `redTeam.contract`,
`runTrace`, `runFinishedNotify`, `runService`, `fixWatch` and
`jobs/aiAgentRunner` — 555 tests, all passing, with no test edits.

Closes #4451
