+++ apps/api/src/services/aiAgents/runLoop.ts
@@ -43,20 +43,11 @@ import { query } from '@anthropic-ai/claude-agent-sdk';
 import type {
   AgentRunVerdict,
   AiAgentKind,
-  AiAgentMode,
   AiAgentPolicy,
-  AiAgentPolicySnapshot,
-  AiAgentRecipients,
   AiAgentRunProfile,
-  AiAgentTriggerKind,
   AiSweepKind,
-  AlertVerdictOutcome,
-  NarrativeOutcome,
-  SweepFindingsOutcome,
-  TicketTriageProposal,
-  TicketTriageSkip,
 } from '@breeze/shared';
-import { AI_AGENT_LIMIT_DEFAULTS, AI_SWEEP_KINDS } from '@breeze/shared';
+import { AI_SWEEP_KINDS } from '@breeze/shared';
 import { envFlag } from '../../config/env';
 import {
   db,
@@ -138,12 +129,35 @@ import { isVerdictProfile, verdictLimits, verdictToolAllowlist } from './verdict
 import { isSweepProfile, sweepLimits, sweepToolAllowlist } from './sweepProfile';
 import { isNarrativeProfile, narrativeLimits, narrativeToolAllowlist } from './narrativeProfile';
 import { isTriageProfile, triageLimits, triageToolAllowlist } from './triageProfile';
-import { loadSweepEvidence, type SweepEvidence } from './sweepEvidence';
-import { loadNarrativeContext, type NarrativeContext } from './narrativeContext';
-import { NarrativePersistConflictError, persistNarrativeReport } from './narrativeReport';
-import { persistAlertVerdict, type AlertVerdictIntentInfo } from './alertVerdicts';
-import { persistSweepFindings, type SweepProposalRecord } from './sweepFindings';
-import { persistTicketTriage } from './ticketTriageFindings';
+import {
+  finalizeNarrative,
+  finalizeSweep,
+  finalizeTicketTriage,
+  finalizeVerdict,
+} from './runFinalizers';
+import type {
+  AgentRow,
+  AgentRunOutcome,
+  LoopResult,
+  OutcomeExecutedAction,
+  OutcomeProposedAction,
+  RunContext,
+  RunRow,
+} from './runLoopTypes';
+
+/**
+ * Re-exported (issue #4451) so every existing importer of the run outcome
+ * shape keeps its current `./runLoop` path after the types moved into
+ * `runLoopTypes.ts`.
+ */
+export type {
+  AgentRunOutcome,
+  OutcomeExecutedAction,
+  OutcomeProposedAction,
+  TicketProposalOutcome,
+} from './runLoopTypes';
+import { loadSweepEvidence } from './sweepEvidence';
+import { loadNarrativeContext } from './narrativeContext';
 
 /** `ai_agent_runs.summary` is `text`, but a reviewer reads the first screen. */
 const RUN_SUMMARY_MAX_CHARS = 2000;
@@ -159,209 +173,6 @@ export const PROPOSAL_RECORDED_TEXT =
   + 'A human will review it. Do not retry this call and do not look for another way to '
   + 'perform it.';
 
-export interface OutcomeProposedAction {
-  tool: string;
-  action?: string;
-  args: Record<string, unknown>;
-  /**
-   * Set for Tier-3 proposals that reached `action_intents`. Presence alone does
-   * NOT mean a human can still approve it — `createActionIntent` commits and
-   * then cancels an intent nobody is eligible to decide. `intentError` is the
-   * authoritative signal, and `run.intent_ids` only ever lists PENDING ones.
-   */
-  intentId?: string;
-  /**
-   * Set instead of a live approval when the intent could not be created, or was
-   * born terminal (`no_eligible_approvers`).
-   */
-  intentError?: string;
-  /**
-   * Set only for an act-mode downgrade-to-propose that carried a concrete
-   * `normalizeTarget` reason (#3826 cheap nonblocking fix) — e.g. a missing
-   * identity field the manifest requires. Absent for an ordinary shadow-mode
-   * proposal and for a drift/cap-exhaustion downgrade, neither of which has
-   * a single call-specific reason to attach.
-   */
-  downgradeReason?: string;
-}
-
-export interface OutcomeExecutedAction {
-  tool: string;
-  action?: string;
-  /**
-   * The real `ai_tool_executions` row id (wave 4a's execution ledger), or
-   * `'(inline)'` when the ledger write itself failed — a ledger failure must
-   * never block the tool call, so the call still executes and is recorded
-   * with the placeholder id (see `executionLedger.ts` and the pre/post hooks
-   * below). Correlation between the pre and post hook is per-tool FIFO order
-   * (same assumption `allowedPending` already made) — genuine per-invocation
-   * ids need SDK hook support the loop doesn't have yet.
-   */
-  executionId: string;
-  result: 'ok' | 'failed';
-  durationMs: number;
-  /**
-   * Act-mode fields (Part B, Task 4) — set ONLY for a call that actually
-   * dispatched through the act branch (an `ActAssetPin` was pinned for it in
-   * the pre-hook; see `actRevalidation.ts`). Absent for every ordinary
-   * Tier-1/2 auto-executed call, exactly like the pre-Part-B behavior.
-   */
-  execution?: 'succeeded' | 'failed' | 'timeout' | 'unknown';
-  verification?: 'passed' | 'failed' | 'inconclusive' | 'skipped';
-  /** Short, human-readable — never a raw tool input/output blob. */
-  verifyDetail?: string;
-  /** The manifest op key (e.g. `manage_services.restart`) — sanitized for the
-   *  finished-run notification; never the raw tool input. */
-  actOpKey?: string;
-  /** Sanitized target identity (service/process name, script/playbook id, or
-   *  a path COUNT) — see `actTargetSummary` (actVerify.ts). Never a full
-   *  path list or tool input/output. */
-  actTargetName?: string;
-}
-
-/**
- * Wave 6 PR 3 (#3828, Task 4) — the model's structured proposal for a
- * ticket-triggered run. Reserved: nothing in this PR populates it (there is no SDK structured-output wiring yet — the
- * model's only output channel is still the free-text summary `driveSdkLoop`
- * already extracts). The field exists now so the outcome shape, the
- * `ai_agent_runs.outcome` jsonb, and `AiAgentRunTicketProposalDto`
- * (`@breeze/shared`) do not change again when a later task wires it up.
- *
- * `notes` are PROPOSED talking points for a human reviewer — see the plan's
- * "no autonomous notes, not even private" design authority. Nothing here is
- * ever the input to a write: shadow mode + the device-less mutation gate
- * together guarantee `manage_tickets` cannot execute for a ticket run
- * regardless of what this field ever holds (`ticketShadowGuardrail.contract.test.ts`).
- *
- * P2-4 (#4191) compile-forward fix: this had zero writers pre-P2-4 (grepped
- * repo-wide), so it is now a type alias onto the shared `TicketTriageProposal`
- * — the real `submit_ticket_proposal` outcome shape — rather than the old
- * ad-hoc `proposedReply`/`proposedStatus`/`proposedPriority` shape, keeping
- * this file and `runTrace.ts`'s projection of it coherent with the DTO.
- *
- * Task A6 (this task) is what actually populates this field: the post-hook
- * (`createAgentRunPostToolUse`'s `submit_ticket_proposal` case) captures the
- * model's validated submission verbatim (no server-owned rebuild, unlike
- * `submit_narrative`) the moment a `triage`-profile run calls its one
- * outcome tool. Turning a populated value into `manage_tickets` intents/
- * `ticket_drafts` rows is still task A8's job (`finishRun`), not this one's.
- */
-export type TicketProposalOutcome = TicketTriageProposal;
-
-export interface AgentRunOutcome {
-  /** The model's `submit_ticket_proposal` submission on a `triage`-profile
-   *  run — see `TicketProposalOutcome`'s docstring. Absent for every
-   *  non-triage run, and for a triage run that never called the tool. */
-  ticketProposal?: TicketProposalOutcome;
-  /**
-   * Follow-up to #4191/#4301 (issue #4462) — per-slot reasons
-   * `finalizeTicketTriage` did not turn part of `ticketProposal` into a
-   * write (e.g. a field below `TICKET_TRIAGE_CONFIDENCE_FLOOR`, a device
-   * already linked). Set at most once, alongside `ticketProposal`, by
-   * `finalizeTicketTriage` from `persistTicketTriage`'s own return value
-   * (`ticketTriageFindings.ts`). Absent when `ticketProposal` is absent, or
-   * when nothing was skipped. `runTrace.ts`'s `mapTicketProposal` carries
-   * this onto the wire DTO's `AiAgentRunTicketProposalDto.skipped`.
-   */
-  ticketTriageSkipped?: TicketTriageSkip[];
-  proposedActions: OutcomeProposedAction[];
-  executedActions: OutcomeExecutedAction[];
-  deniedActions: Array<{ tool: string; reason: string }>;
-  /** Tool calls that actually EXECUTED (denials and proposals excluded). */
-  toolExecutionCount: number;
-  budgetExceeded?: boolean;
-  wallClockExceeded?: boolean;
-  /** The SDK stopped because `maxTurns` was reached (`error_max_turns`). */
-  maxTurnsExceeded?: boolean;
-  /**
-   * Run-level rollup over every act execution's (execution, verification)
-   * pair, computed once at finish by `computeRunVerdict` — see there for the
-   * exact rule. Absent (rather than defaulted) for anything that never
-   * reaches the rollup — kept optional so old rows read back without it
-   * don't need a migration.
-   */
-  runVerdict?: AgentRunVerdict;
-  /**
-   * Phase 2 wave P2-1 (alert verdicts) — the validated `submit_alert_verdict`
-   * input, captured by the post-tool-use hook on a verdict-profile run. Set
-   * at most once (the outcome tool's own description tells the model to call
-   * it exactly once); absent for a `full`-profile run, which never has the
-   * outcome tool exposed at all. Persisted to `ai_alert_verdicts` by
-   * `finalizeVerdict` (below), which also carries this value into
-   * `finishRun`'s persisted `outcome` jsonb.
-   */
-  alertVerdict?: AlertVerdictOutcome;
-  /**
-   * Phase 2 wave P2-1 (alert verdicts), review round 1 (IMPORTANT 2) — the
-   * disposition of `alertVerdict.suggestedAction`'s Tier-2 intent attempt,
-   * set by `finalizeVerdict` alongside `alertVerdict` itself. Absent when
-   * there was no `suggestedAction` to act on in the first place (nothing to
-   * report) OR for a `full`-profile run. Display strings only — NEVER the
-   * raw `Error.message` from `createActionIntent` (that could carry
-   * tool-input detail); see `AlertVerdictIntentInfo`'s own docstring.
-   */
-  alertVerdictIntent?: AlertVerdictIntentInfo;
-  /**
-   * Phase 2 wave P2-2 (scheduled sweeps) — the validated
-   * `submit_sweep_findings` input, captured by the post-tool-use hook on a
-   * sweep-profile run. Set at most once (the outcome tool's own description
-   * and the sweep task turn both tell the model to call it exactly once);
-   * absent for every other profile, which never has the sweep outcome tool
-   * exposed at all. Task A7 persists it and turns each accepted
-   * `proposedAction` into a supervised action intent — NOTHING here executes.
-   */
-  sweepFindings?: SweepFindingsOutcome;
-  /**
-   * Phase 2 wave P2-2 (scheduled sweeps), Task A7 — one bookkeeping record
-   * per finding that carried a `proposedAction`, written by `finalizeSweep`
-   * below. Absent when the run proposed nothing at all (nothing to report on)
-   * and for every non-sweep profile. Display strings only — NEVER the raw
-   * `Error.message` from `createActionIntent`; see `SweepProposalRecord`'s own
-   * docstring (sweepFindings.ts).
-   */
-  sweepProposals?: SweepProposalRecord[];
-  /**
-   * Phase 2 wave P2-2, Task A7 — whether the SYSTEM evidence this sweep run
-   * reported on was capped/byte-trimmed (`SweepEvidence.truncated`, Task 5).
-   * Persisted here because the evidence itself is never stored on the run,
-   * and the run-detail projection has to be able to tell a reader "the model
-   * saw a sample, not the whole fleet". Absent for every non-sweep profile.
-   */
-  sweepEvidenceTruncated?: boolean;
-  /**
-   * Phase 2 wave P2-3 (weekly org narrative) — the weekly narrative, captured
-   * by the post-tool-use hook on a narrative-profile run. Set at most once
-   * (the outcome tool's description and the narrative task turn both tell the
-   * model to call it exactly once); absent for every other profile, which
-   * never has the narrative outcome tool exposed at all.
-   *
-   * NOT the raw tool input, unlike its two siblings above: the model submits
-   * `{ headline, sections: [{ key, bullets }] }` and the SERVER attaches the
-   * section titles, imposes the canonical section order and derives the
-   * markdown (`narrativeOutcomeFromSubmission`, reached through
-   * `validateOutcomeToolInput`). See `orgNarrativeReport.ts`'s file docstring
-   * for why the model never authors any of those three.
-   *
-   * Task A7 persists it as a system-authored report artifact and links
-   * `ai_agent_runs.report_run_id` — NOTHING here executes.
-   */
-  narrative?: NarrativeOutcome;
-  /**
-   * Phase 2 wave P2-3, Task A7 — the system-authored report the narrative was
-   * materialised into, written by `finalizeNarrative` once
-   * `persistNarrativeReport`'s transaction committed. Absent for every other
-   * profile, for a narrative run that produced nothing, and for one whose
-   * persistence lost the CAS (in which case the run's `error_code` says so).
-   *
-   * TWO ids and nothing else. `runFinishedNotify` reads exactly this to build
-   * the notification's `metadata.narrative`, and the run-detail route reads
-   * `ai_agent_runs.report_run_id` (the typed column) rather than this jsonb —
-   * so nothing downstream is tempted to grow a copy of the narrative, let
-   * alone of the weekly context, on the run row.
-   */
-  narrativeReport?: { reportId: string; reportRunId: string };
-}
-
 /** Error carrying the short code that lands in `ai_agent_runs.error_code` (varchar(64)). */
 export class AgentRunError extends Error {
   readonly errorCode: string;
@@ -371,117 +182,6 @@ export class AgentRunError extends Error {
     this.name = 'AgentRunError';
     this.errorCode = errorCode;
   }
-}
-
-interface RunRow {
-  id: string;
-  agentId: string;
-  orgId: string;
-  deviceId: string | null;
-  alertId: string | null;
-  /** Wave 6 PR 3 (#3828, Task 1/4) — the triggering ticket for `triggerKind:
-   *  'ticket'` runs; always `null` for every other trigger kind. */
-  ticketId: string | null;
-  /** Wave 6 PR 4 (#3828, Task 1/4) — the triggering canonical incident for
-   *  `triggerKind: 'anomaly'` runs; always `null` for every other trigger kind. */
-  anomalyIncidentId: string | null;
-  status: string;
-  modeAtStart: Exclude<AiAgentMode, 'off'>;
-  triggerKind: AiAgentTriggerKind;
-  policySnapshot: AiAgentPolicySnapshot;
-  /** Phase 2 wave P2-1 (alert verdicts) — see `verdictProfile.ts`. */
-  profile: AiAgentRunProfile;
-  correlationGroupId: string | null;
-  /** Phase 2 wave P2-2 (scheduled sweeps) — the `ai_agent_schedules` row a
-   *  `sweep`-profile run was fanned out from; `null` for every other trigger. */
-  scheduleId: string | null;
-  /** Free-form trigger provenance (`jsonb`, defaults `{}`). A sweep run's
-   *  carries `{ scheduleId, occurrenceKey, sweepKinds }`, written by the
-   *  fixed-tick sweeper — read DEFENSIVELY below (any field may be missing or
-   *  the wrong shape; the column has no compile-time schema). */
-  triggerRef: Record<string, unknown>;
-}
-
-interface AgentRow {
-  id: string;
-  orgId: string | null;
-  partnerId: string | null;
-  name: string;
-  kind: AiAgentKind;
-  recipients: Partial<AiAgentRecipients>;
-}
-
-interface RunContext {
-  run: RunRow;
-  agent: AgentRow;
-  orgPartnerId: string;
-  device: { id: string; siteId: string; hostname: string; osType: string } | null;
-  alert: { title: string; severity: string; message: string | null } | null;
-  /** Bounded, sanitized ticket context (Task 4) — `null` for every non-ticket
-   *  run, and for a ticket run whose ticket has vanished/moved org (same
-   *  "moved/deleted reads as absent" posture `device`/`alert` already use). */
-  ticket: TicketRunContext | null;
-  /** Bounded anomaly context (wave 6 PR 4, #3828, Task 4) — `null` for every
-   *  non-anomaly run, and for an anomaly run whose incident has vanished/
-   *  moved org (same "moved/deleted reads as absent" posture as `ticket`
-   *  above). */
-  anomaly: AnomalyRunContext | null;
-  /**
-   * Phase 2 wave P2-1 (alert verdicts). Set only when `run.correlationGroupId`
-   * is set AND the group still resolves inside the run's org — same
-   * missing-reads-as-null shape as `device`/`alert` above (a group can be
-   * deleted/re-scoped between admission and delivery). `correlationTypes` is
-   * the `metadata.correlationTypes` array, narrowed to strings only — the
-   * jsonb column carries no compile-time shape.
-   */
-  correlationGroup: {
-    id: string;
-    memberCount: number;
-    noiseReductionPercent: number;
-    rootAlertId: string | null;
-    correlationTypes: string[];
-  } | null;
-  /**
-   * Phase 2 wave P2-2 (scheduled sweeps) — the schedule occurrence and the
-   * bounded, system-collected evidence this run reports on. Set only for a
-   * `sweep`-profile run; `null` for every other profile. An empty
-   * `kinds`/`evidence` is a legitimate value (see the loader below), never a
-   * reason to fail the run.
-   */
-  sweep: {
-    scheduleId: string;
-    occurrenceKey: string;
-    kinds: AiSweepKind[];
-    evidence: SweepEvidence;
-  } | null;
-  /**
-   * Phase 2 wave P2-3 (weekly org narrative) — the schedule occurrence and the
-   * bounded, system-assembled week of activity this run writes about. Set only
-   * for a `narrative`-profile run; `null` for every other profile.
-   *
-   * A narrative run loads NO sweep evidence and NO device context: it is
-   * device-less by construction and its tool floor is empty, so this context
-   * is its entire input. An empty/unavailable block inside it is a legitimate
-   * value the prompt renders as "(not measured)", never a reason to fail the
-   * run.
-   *
-   * `scheduleId` is `''` when the run carries none — Task A7's
-   * `finalizeNarrative` reports `narrative_no_schedule` rather than persisting
-   * an artifact it cannot attribute to a schedule.
-   */
-  narrative: {
-    scheduleId: string;
-    occurrenceKey: string;
-    context: NarrativeContext;
-  } | null;
-  /**
-   * The execution-ledger `ai_sessions` row for this run (Task 1/2). Set once,
-   * inside `driveSdkLoop`, right after the model is resolved — `null` until
-   * then, and stays `null` for the lifetime of the run if session creation
-   * itself failed (a best-effort write: see `driveSdkLoop`). `finishRun` reads
-   * it back to reconcile/close the session.
-   */
-  sessionId: string | null;
 }
 
 /**
@@ -1517,46 +1217,6 @@ function promptContext(ctx: RunContext, effective: AiAgentPolicy): AgentRunPromp
   };
 }
 
-interface LoopResult {
-  summary: string;
-  costCents: number;
-  turnCount: number;
-  outcome: AgentRunOutcome;
-  intentIds: string[];
-  /**
-   * The run's `ai_agent` principal context (built once, near the top of
-   * `driveSdkLoop`, via `buildAgentAuthContext`) — carried out on `result`
-   * rather than recomputed, since `finishRun` (Task 8, P2-1) needs it to call
-   * `persistAlertVerdict`/`createActionIntent` for a verdict run and has no
-   * other way to reach it: `driveSdkLoop` has exactly one return statement,
-   * reached only after `buildAgentAuthContext` has already succeeded (a
-   * throw there returns straight to `executeAgentRun`'s catch block, never
-   * through here — see that call site's own comment), so this is always
-   * populated whenever `finishRun` reads it. Smaller diff than widening
-   * `RunContext` with a second mutated-after-construction field alongside
-   * `sessionId`.
-   */
-  agentAuth: AuthContext;
-  /**
-   * Set when the SDK loop itself threw. Carried back rather than rethrown so
-   * the tokens already burned still land on the run row — a crashed run that
-   * recorded `cost_cents: 0` would make the agent's daily budget cap
-   * under-count real spend. Setup failures BEFORE any spend still throw.
-   */
-  failure?: { errorCode: string; message: string };
-  /**
-   * Phase 2 wave P2-4 (#4191), Task A8 — the SUBSET of `intentIds` that a
-   * creation-time `ticket_autonomy` grant already decided (status
-   * `approved`, no human fan-out ever happened — see `ticketAutonomy.ts`).
-   * Populated only by `finalizeTicketTriage`; every other profile's
-   * finalizer only ever links a `pending_approval` intent, so this stays
-   * empty (and every existing `intentIds.length > 0` reader is unaffected)
-   * for `verdict`/`sweep`/`narrative`/`full` runs. See
-   * `classifyIntentAwaitingApproval`'s own docstring for how this is used.
-   */
-  decidedIntentIds?: string[];
-}
-
 async function driveSdkLoop(ctx: RunContext, effective: AiAgentPolicy): Promise<LoopResult> {
   const { run } = ctx;
   const limits = effective.limits;
@@ -2129,363 +1789,6 @@ const TERMINAL_EVENT = {
   awaiting_approval: 'ai.agent.run.awaiting_approval',
   failed: 'ai.agent.run.failed',
 } as const;
-
-/**
- * Phase 2 wave P2-1 (alert verdicts), Task 8 — review round 1 fixes
- * (IMPORTANT 3 + 4). Runs from `executeAgentRun`, BEFORE the
- * awaiting_approval/completed decision and before any of the three
- * `finishRun` call sites — not inside `finishRun` itself (moved there in
- * the original Task 8 pass; the ordering was the bug IMPORTANT 3 found: the
- * status ternary read `intentIds.length` before this ran, so a verdict run
- * that created a pending intent still finished `completed`).
- *
- * Returns an errorCode override. The caller applies it ONLY to the
- * normal-finish `finishRun` call — the `result.failure`/ceiling branches
- * already carry a real, more specific code (`llm_unavailable`,
- * `max_turns_exceeded`, …) that must not be clobbered with the generic
- * `verdict_missing`/`verdict_persist_failed`, and `classifyTerminal`
- * (agentCircuit.ts) still needs that real code to classify a genuine
- * runner failure correctly for the circuit breaker regardless of profile.
- * Returning `null` and letting those two branches ignore it entirely keeps
- * that scoping intact without this function needing to know which branch
- * the caller is about to take.
- */
-async function finalizeVerdict(ctx: RunContext, result: LoopResult): Promise<string | null> {
-  if (!isVerdictProfile(ctx.run)) return null;
-  const { outcome } = result;
-
-  if (!outcome.alertVerdict) {
-    // A verdict run that reached a normal finish without ever calling
-    // `submit_alert_verdict` is a runner failure a human needs to see —
-    // never a silent `completed` with nothing to show for it. (On the
-    // failure/ceiling branches this return value is discarded by the
-    // caller, so this only actually takes effect on the normal-finish path
-    // — see this function's own docstring.)
-    outcome.runVerdict = 'needs_attention';
-    return 'verdict_missing';
-  }
-
-  // IMPORTANT 4: re-read the run's live status right before writing — the
-  // stall reaper (`reapStalledAgentRuns`, runService.ts) or a second
-  // executor may have already moved this run out of `running` while the
-  // SDK loop was in flight. Skip persistence rather than orphan a live
-  // verdict row + a live approval request under a run nobody owns anymore.
-  // A tiny window remains between this read and the writes inside
-  // `persistAlertVerdict` below — accepted per review ruling.
-  const stillRunning = await isRunStillRunning(ctx.run.id, ctx.run.orgId);
-  if (!stillRunning) {
-    console.warn('[aiAgentRunLoop] skipped verdict persistence — run left `running` before it could be persisted', {
-      runId: ctx.run.id,
-    });
-    return null;
-  }
-
-  try {
-    const { intentId, suggestionDisposition, suggestionReason } = await persistAlertVerdict(
-      {
-        id: ctx.run.id,
-        orgId: ctx.run.orgId,
-        alertId: ctx.run.alertId,
-        correlationGroupId: ctx.run.correlationGroupId,
-        deviceId: ctx.run.deviceId,
-        // Review round 2 (IMPORTANT 1) — the run's OWN effective
-        // toolAllowlist, off the already-loaded run row's policySnapshot
-        // (never re-queried): `persistAlertVerdict` gates a suggested
-        // mutation's intent creation on this SAME authority the release
-        // path re-checks (agentReleaseAuthority.ts).
-        toolAllowlist: ctx.run.policySnapshot.effective.toolAllowlist,
-      },
-      outcome.alertVerdict,
-      result.agentAuth,
-    );
-    if (intentId) result.intentIds.push(intentId);
-    // Review round 1 (IMPORTANT 2): only recorded when there was a
-    // `suggestedAction` to report on at all — see `AlertVerdictIntentInfo`'s
-    // own docstring for why this is absent rather than a vacuous
-    // `not_created` on every verdict.
-    if (outcome.alertVerdict.suggestedAction) {
-      const intentInfo: AlertVerdictIntentInfo = { disposition: suggestionDisposition };
-      if (suggestionReason) intentInfo.reason = suggestionReason;
-      outcome.alertVerdictIntent = intentInfo;
-    }
-    return null;
-  } catch (error) {
-    console.error('[aiAgentRunLoop] failed to persist alert verdict', { runId: ctx.run.id, error });
-    return 'verdict_persist_failed';
-  }
-}
-
-/**
- * Phase 2 wave P2-2 (scheduled sweeps), Task A7 — the sweep sibling of
- * `finalizeVerdict` above, called from the SAME place in `executeAgentRun`
- * (before the awaiting_approval/completed decision, outside any ambient DB
- * context — `persistSweepFindings` inherits `createActionIntent`'s
- * no-ambient-context precondition; see `alertVerdicts.ts:208-226`).
- *
- * Returns an errorCode override the caller applies ONLY on the normal-finish
- * path, exactly as it does for `finalizeVerdict`'s. Unlike a verdict run, a
- * sweep run that produced no findings is NOT an error: task 6 already treats
- * `outcome.sweepFindings` as this profile's "produced something" signal, and a
- * sweep that legitimately found nothing to report still completes cleanly.
- *
- * `evidenceDeviceIds` is built from `ctx.sweep.evidence` — the rows the
- * SYSTEM loaded, not anything the model said — and is the control that keeps
- * a crafted proposal from reaching a device this run never looked at. A run
- * whose context carries no sweep block at all (defensive: `ctx.sweep` is
- * populated for every sweep-profile run today) therefore has an EMPTY
- * evidence set, which refuses every proposal — fail closed, never open.
- */
-async function finalizeSweep(ctx: RunContext, result: LoopResult): Promise<string | null> {
-  if (!isSweepProfile(ctx.run)) return null;
-  const { outcome } = result;
-
-  // Recorded even when there are no findings: "the model only saw a sample"
-  // is true of the run regardless of what it chose to report.
-  if (ctx.sweep) outcome.sweepEvidenceTruncated = ctx.sweep.evidence.truncated;
-
-  if (!outcome.sweepFindings) return null;
-
-  // Same IMPORTANT-4 re-read as `finalizeVerdict`: the stall reaper
-  // (`reapStalledAgentRuns`) or a second executor may have moved this run out
-  // of `running` while the SDK loop was in flight. Skip persistence rather
-  // than mint live, human-approvable intents under a run nobody owns anymore.
-  const stillRunning = await isRunStillRunning(ctx.run.id, ctx.run.orgId);
-  if (!stillRunning) {
-    console.warn('[aiAgentRunLoop] skipped sweep persistence — run left `running` before it could be persisted', {
-      runId: ctx.run.id,
-    });
-    return null;
-  }
-
-  const evidenceDeviceIds = new Set<string>();
-  for (const kind of Object.values(ctx.sweep?.evidence.kinds ?? {})) {
-    for (const row of kind?.rows ?? []) {
-      if (row.deviceId) evidenceDeviceIds.add(row.deviceId);
-    }
-  }
-
-  try {
-    const { proposals, intentIds } = await persistSweepFindings(
-      {
-        id: ctx.run.id,
-        orgId: ctx.run.orgId,
-        agentId: ctx.run.agentId,
-        // A sweep run is device-less by construction — that is exactly why
-        // every intent it mints carries its own explicit device scope.
-        deviceId: null,
-        scheduleId: ctx.run.scheduleId,
-        // The AGENT's effective allowlist off the already-loaded run row —
-        // the same authority `agentReleaseAuthority.ts` re-checks at release.
-        toolAllowlist: ctx.run.policySnapshot.effective.toolAllowlist,
-        // The AGENT's action cap, NOT `sweepLimits`' hard `0` (which governs
-        // what the run LOOP may execute, and a sweep executes nothing). `??`
-        // tolerates a v1 policy snapshot, which predates the field entirely.
-        maxActionsPerRun:
-          ctx.run.policySnapshot.effective.limits.maxActionsPerRun
-          ?? AI_AGENT_LIMIT_DEFAULTS.maxActionsPerRun,
-        evidenceDeviceIds,
-      },
-      outcome.sweepFindings,
-      result.agentAuth,
-    );
-    if (proposals.length > 0) outcome.sweepProposals = proposals;
-    for (const intentId of intentIds) result.intentIds.push(intentId);
-    return null;
-  } catch (error) {
-    console.error('[aiAgentRunLoop] failed to persist sweep findings', { runId: ctx.run.id, error });
-    return 'sweep_persist_failed';
-  }
-}
-
-/**
- * Phase 2 wave P2-3 (weekly org narrative), Task A7 — the narrative sibling of
- * `finalizeVerdict`/`finalizeSweep` above, called from the SAME place in
- * `executeAgentRun` and outside any ambient DB context (`persistNarrativeReport`
- * opens its own system transaction and would take a second pooled connection
- * from inside one).
- *
- * Returns an errorCode override the caller applies ONLY on the normal-finish
- * path, exactly as it does for the two siblings'.
- *
- * A narrative run that produced NOTHING is an error, unlike a sweep that found
- * nothing: a sweep legitimately reports "all clear", but the narrative IS the
- * deliverable — an occurrence that yields no document is a silent hole in a
- * weekly series an MSP forwards to their customer, and `narrative_missing` is
- * what makes that visible on the run row instead of a clean `completed` with
- * nothing behind it.
- *
- * `narrative_no_schedule` is deliberately separate from a persistence failure:
- * the definition's identity IS `(org_id, source_ai_agent_schedule_id)`, so
- * without a schedule there is no row to find-or-create and no idempotency at
- * all — a manually triggered narrative run would mint a fresh definition every
- * time. Refusing is the conservative answer; the narrative still lives on the
- * run's own outcome and renders on the run-detail page.
- */
-async function finalizeNarrative(ctx: RunContext, result: LoopResult): Promise<string | null> {
-  if (!isNarrativeProfile(ctx.run)) return null;
-  const { outcome } = result;
-
-  if (!outcome.narrative) return 'narrative_missing';
-  if (!ctx.narrative?.scheduleId) {
-    console.warn('[aiAgentRunLoop] narrative run carries no schedule — not persisting a report definition', {
-      runId: ctx.run.id, orgId: ctx.run.orgId,
-    });
-    return 'narrative_no_schedule';
-  }
-
-  // Same IMPORTANT-4 re-read both siblings carry: the stall reaper
-  // (`reapStalledAgentRuns`) or a second executor may have moved this run out
-  // of `running` while the SDK loop was in flight. Skip rather than publish a
-  // customer-facing artifact under a run nobody owns any more. (A tiny window
-  // remains between this read and the transaction below — closed properly
-  // there, by the `FOR UPDATE` lock plus the `report_run_id IS NULL` CAS.)
-  const stillRunning = await isRunStillRunning(ctx.run.id, ctx.run.orgId);
-  if (!stillRunning) {
-    console.warn('[aiAgentRunLoop] skipped narrative persistence — run left `running` before it could be persisted', {
-      runId: ctx.run.id,
-    });
-    return null;
-  }
-
-  try {
-    const { reportId, reportRunId } = await persistNarrativeReport({
-      run: {
-        id: ctx.run.id,
-        orgId: ctx.run.orgId,
-        agentId: ctx.run.agentId,
-        scheduleId: ctx.narrative.scheduleId,
-      },
-      agent: { id: ctx.agent.id, name: ctx.agent.name },
-      occurrenceKey: ctx.narrative.occurrenceKey || null,
-      context: ctx.narrative.context,
-      outcome: outcome.narrative,
-    });
-    // TWO ids, never the narrative or the context — see the field's docstring.
-    outcome.narrativeReport = { reportId, reportRunId };
-    return null;
-  } catch (error) {
-    if (error instanceof NarrativePersistConflictError) {
-      // A race that resolved correctly (another executor linked an artifact,
-      // or the run moved) — logged at warn, not error, and given its own code
-      // so it is distinguishable from a genuine write failure on the run row.
-      console.warn('[aiAgentRunLoop] narrative artifact was not linked to this run', {
-        runId: ctx.run.id, reason: (error as Error).message,
-      });
-      return 'narrative_persist_conflict';
-    }
-    console.error('[aiAgentRunLoop] failed to persist the narrative report', { runId: ctx.run.id, error });
-    return 'narrative_persist_failed';
-  }
-}
-
-/**
- * Phase 2 wave P2-4 (ticket triage, #4191), Task A8 — the triage sibling of
- * `finalizeVerdict`/`finalizeSweep`/`finalizeNarrative` above, called from
- * the SAME place in `executeAgentRun` and outside any ambient DB context
- * (`persistTicketTriage` inherits `createActionIntent`'s no-ambient-context
- * precondition; see `sweepFindings.ts`/`alertVerdicts.ts` headers).
- *
- * Returns an errorCode override the caller applies ONLY on the
- * normal-finish path, exactly as its three siblings do.
- *
- * A triage run that reached a normal finish without ever calling
- * `submit_ticket_proposal` is treated like a verdict run's missing verdict,
- * not like a sweep's legitimate "found nothing": the prompt tells the model
- * to call its one outcome tool exactly once (`runnerPrompt.ts`), so a run
- * that never did is a runner failure a human needs to see, not a silent
- * `completed`.
- */
-async function finalizeTicketTriage(ctx: RunContext, result: LoopResult): Promise<string | null> {
-  if (!isTriageProfile(ctx.run)) return null;
-  const { outcome } = result;
-
-  if (!outcome.ticketProposal) return 'ticket_proposal_missing';
-
-  // Defensive, mirroring `finalizeNarrative`'s `ctx.narrative?.scheduleId`
-  // check: a `triage`-profile run is admitted with `triggerKind: 'ticket'`
-  // and a `ticketId` by construction (wave 6 PR 3), but this is the one
-  // place a violated invariant would otherwise NPE deep inside
-  // `persistTicketTriage` instead of surfacing a clear error code.
-  if (ctx.run.triggerKind !== 'ticket' || !ctx.run.ticketId) {
-    console.warn('[aiAgentRunLoop] triage run carries no ticket scope — not persisting its proposal', {
-      runId: ctx.run.id, orgId: ctx.run.orgId, triggerKind: ctx.run.triggerKind,
-    });
-    return 'ticket_triage_scope_missing';
-  }
-
-  // Same IMPORTANT-4 re-read every sibling carries: the stall reaper
-  // (`reapStalledAgentRuns`) or a second executor may have moved this run
-  // out of `running` while the SDK loop was in flight. Skip persistence
-  // rather than mint live, human-or-autonomy-decided intents under a run
-  // nobody owns anymore.
-  const stillRunning = await isRunStillRunning(ctx.run.id, ctx.run.orgId);
-  if (!stillRunning) {
-    console.warn('[aiAgentRunLoop] skipped ticket-triage persistence — run left `running` before it could be persisted', {
-      runId: ctx.run.id,
-    });
-    return null;
-  }
-
-  try {
-    const { intentIds, approvedIntentIds, skipped } = await persistTicketTriage(
-      {
-        id: ctx.run.id,
-        orgId: ctx.run.orgId,
-        agentId: ctx.run.agentId,
-        ticketId: ctx.run.ticketId,
-        // The run's IMMUTABLE start-of-run snapshot — gate 2 of the
-        // advisory autonomy check (see `ticketTriageFindings.ts`'s header).
-        policySnapshot: ctx.run.policySnapshot,
-        // The AGENT's action cap — a triage run's ONE deliberate divergence
-        // from its siblings' hard-zeroed `maxActionsPerRun` (see
-        // `triageLimits`'s docstring): this is a POST-run minting cap, not
-        // an in-run tool-call budget. `??` tolerates a pre-v8 snapshot.
-        maxActionsPerRun:
-          ctx.run.policySnapshot.effective.limits.maxActionsPerRun
-          ?? AI_AGENT_LIMIT_DEFAULTS.maxActionsPerRun,
-      },
-      outcome.ticketProposal,
-      result.agentAuth,
-    );
-    for (const intentId of intentIds) result.intentIds.push(intentId);
-    // GROUND TRUTH, not the call-level `autonomous` advisory flag (review
-    // fix, round 2): `approvedIntentIds` is built per-intent from each
-    // intent's OWN returned status inside `persistTicketTriage` — the live
-    // gates can flip between two sequential `createActionIntent` calls in
-    // that same invocation, so a uniform "autonomous ⇒ every id is decided"
-    // read would wrongly classify a run where one intent lands `approved`
-    // and the next lands `pending_approval`. See
-    // `classifyIntentAwaitingApproval`'s docstring for how this is used.
-    if (approvedIntentIds.length > 0) {
-      result.decidedIntentIds = [...(result.decidedIntentIds ?? []), ...approvedIntentIds];
-    }
-    if (skipped.length > 0) {
-      // Issue #4462: previously logged only — never reached `result.outcome`,
-      // so `finishRun`'s persisted `outcome` jsonb (and everything derived
-      // from it, including the run DTO) never carried it. Same direct-mutate
-      // pattern as `outcome.alertVerdictIntent` above.
-      outcome.ticketTriageSkipped = skipped;
-      console.info('[aiAgentRunLoop] ticket-triage proposal partially skipped', {
-        runId: ctx.run.id, skipped,
-      });
-    }
-    return null;
-  } catch (error) {
-    console.error('[aiAgentRunLoop] failed to persist the ticket-triage proposal', { runId: ctx.run.id, error });
-    return 'ticket_triage_persist_failed';
-  }
-}
-
-/** IMPORTANT 4 helper — see `finalizeVerdict`. System-scoped: this runs from
- *  the background run loop, with no ambient request context. */
-async function isRunStillRunning(runId: string, orgId: string): Promise<boolean> {
-  return inSystemDbContext(async () => {
-    const [row] = await db.select({ status: aiAgentRuns.status }).from(aiAgentRuns)
-      .where(and(eq(aiAgentRuns.id, runId), eq(aiAgentRuns.orgId, orgId)))
-      .limit(1);
-    return row?.status === 'running';
-  });
-}
 
 /**
  * Act-execution op evidence (Task 6, P2-5, #4192) — one `executed`/`failed`
